First cleaning in SR initstr: removal of cb55
This cb was filled for time=tstart and time=tstart + 30 min
  - the variables sk0, sl0 and dtrad0 were filled at time=tstart
  - the variables sk1, sl1 and dtrad1 were filled at time=tstart + 30 min.

Note that the variables time0 and time2 (the logic would have named it time1...) were both equal to tstart + 30 min. Mistake?
Similarly, the indexes 1 of dtrad0 and dtrad1 were undefined.

This cb was then written in restart files (in SR outm, in file outp.f). In case of a restart, the variables stored in cb55 were read from the restart file (in SR startm, in file str.f), but not used.